% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anndataToSeurat.R
\name{anndataToSeurat}
\alias{anndataToSeurat}
\title{anndataToSeurat}
\usage{
anndataToSeurat(adata_path, image = FALSE)
}
\arguments{
\item{adata_path}{String file path to AnnData H5AD Object.}

\item{image}{Logical Import H & E/IHC Image. FALSE by default.}
}
\value{
Seurat Object
}
\description{
Reads in an AnnData H5AD object from a provided file path, and converts to a Seurat Object. Compatible with single-cell and spatial objects, including with and without H & E and immunofluorescent images.
}
