% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipImageSeurat.R
\name{flipImageSeurat}
\alias{flipImageSeurat}
\title{flipImageSeurat}
\usage{
flipImageSeurat(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat Object with immunofluorescent images.}
}
\value{
Seurat Object with flipped images.
}
\description{
When importing immunofluorescent images from ScanPy into Seurat, the images may require flipping, which can be done using this function.
If needed for SpatialExperiment objects, use existing SpatialExperiment::mirrorObject function.
}
