% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratToAnndata.R
\name{seuratToAnndata}
\alias{seuratToAnndata}
\title{seuratToAnndata}
\usage{
seuratToAnndata(
  seurat_obj,
  export_path,
  raw_counts_names = c("RNA", "counts"),
  image = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat Object.}

\item{export_path}{String file path to export file to.}

\item{raw_counts_names}{Vector c(assay_name, layer_name) for assay layer to be used as raw counts matrix 'X' by AnnData. c('RNA', 'counts') by default.}

\item{image}{Logical Import H & E/IHC Image. FALSE by default.}
}
\value{
None
}
\description{
Provided a Seurat object, converts to an Anndata object, which is exported as a h5ad file to a provided filepath.Compatible with single-cell and spatial objects, including with and without H & E and immunofluorescent images.
}
