% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratToSpatialExperiment.R
\name{seuratToSpatialExperiment}
\alias{seuratToSpatialExperiment}
\title{seuratToSpatialExperiment}
\usage{
seuratToSpatialExperiment(seurat_obj, image = FALSE)
}
\arguments{
\item{seurat_obj}{Seurat Object.}

\item{image}{Logical Import H & E/IHC Image. FALSE by default.}
}
\value{
SingleCellExperiment or SpatialExperiment Object
}
\description{
Provided a Seurat object, converts to a SpatialExperiment/SingleCellExperiment Object. Compatible with single-cell and spatial objects, including with and without H & E and immunofluorescent images.
}
