% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialExperimentToAnndata.R
\name{spatialExperimentToAnndata}
\alias{spatialExperimentToAnndata}
\title{spatialExperimentToAnndata}
\usage{
spatialExperimentToAnndata(
  spe,
  export_path,
  raw_counts_name = "counts",
  image = FALSE
)
}
\arguments{
\item{spe}{SingleCellExperiment/SpatialExperiment Object}

\item{export_path}{String file path to export file to.}

\item{raw_counts_name}{String Name of assay with unnormalised counts data, to be used as raw counts matrix 'X' by AnnData. "counts" by default.}

\item{image}{Logical Import H & E/IHC Image. FALSE by default.}
}
\value{
None
}
\description{
Provided a SingleCellExperiment/SpatialExperiment Object, converts to an AnnData object, which is exported as a h5ad file to a provided filepath. Compatible with single-cell and spatial objects, including with and without H & E and immunofluorescent images.
}
