% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialExperimentToSeurat.R
\name{spatialExperimentToSeurat}
\alias{spatialExperimentToSeurat}
\title{spatialExperimentToSeurat}
\usage{
spatialExperimentToSeurat(spe, raw_counts_name = "counts", image = FALSE)
}
\arguments{
\item{spe}{SingleCellExperiment/SpatialExperiment Object}

\item{raw_counts_name}{String Name of assay with unnormalised counts data, to be provided as assay counts. "counts" by default.}

\item{image}{Logical Import H & E/IHC Image. FALSE by default.}
}
\value{
Seurat Object.
}
\description{
Provided a SingleCellExperiment/SpatialExperiment Object, converts it to a Seurat Object. Compatible with single-cell and spatial objects, including with and without H & E and immunofluorescent images.
}
